requirements.txt
streamlit
pandas
curl_cffi
requests
beautifulsoup4
lxml
duckduckgo-search
googlesearch-python
